apps/api/pyproject.toml
[project]
name = "rpcp-api"
version = "0.1.0"
description = "Revenue & Procurement Copilot Platform API"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.6",
  "sqlalchemy>=2.0.35",
  "alembic>=1.13.2",
  "pydantic>=2.9.2",
  "pydantic-settings>=2.5.2",
  "python-multipart>=0.0.9",
  "email-validator>=2.2.0",
  "passlib[bcrypt]>=1.7.4",
  "pyjwt>=2.9.0",
  "httpx>=0.27.2",
  "pgvector>=0.3.4",
  "psycopg[binary]>=3.2.1",
  "orjson>=3.10.7",
  "tenacity>=9.0.0",
  "structlog>=24.4.0",
  "python-dateutil>=2.9.0.post0",
  "greenlet>=3.1.1",
  "pypdf>=5.1.0",
  "python-docx>=1.1.2"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3",
  "pytest-asyncio>=0.24.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.6.8",
  "mypy>=1.11.2"
]

[build-system]
requires = ["setuptools>=75.1.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
